Code cleanup and write of the importmaps api endpoint
