+++ src/api/apps.js
@@ -1,9 +1,9 @@
 'use strict'
 
 import logger from 'winston'
-
 import * as authorisation from './authorisation'
 import {AppModelAPI} from '../model/apps'
+import {DEFAULT_IMPORT_MAP_PATHS} from '../constants'
 
 /*
   Checks admin permission for create, update and delete operations.
@@ -149,5 +149,31 @@ export async function deleteApp(ctx, appId) {
     })
   } catch (e) {
     createErrorResponse(ctx, 'delete', e)
+  }
+}
+
+/**
+ * Retrieves all apps from the database and transforms the data into an enriched import map json response
+ * @param {*} ctx
+ */
+export async function getTransformedImportMap(ctx) {
+  try {
+    const importMaps = await AppModelAPI.find(ctx.request.query, 'name url')
+
+    logger.info(
+      `Fetched ${importMaps.length} apps for importmaps`
+    )
+
+    const imports = importMaps.reduce((acc, curr) => {
+      acc[curr.name] = curr.url
+      return acc
+    }, {})
+
+    const mergedImports = {...DEFAULT_IMPORT_MAP_PATHS, ...imports}
+
+    ctx.body = {imports: mergedImports}
+    ctx.status = 200
+  } catch (e) {
+    logger.error(`Could not retrieve an enriched importamap via the API: ${e}`)
   }
 }
