+++ src/model/apps.js
@@ -3,6 +3,8 @@
 import {Schema} from 'mongoose'
 
 import {connectionAPI, connectionDefault} from '../config'
+import { loggers } from 'winston'
+import axios from 'axios'
 
 const AppSchema = new Schema({
   name: {
@@ -32,3 +34,21 @@ const AppSchema = new Schema({
 
 export const AppModelAPI = connectionAPI.model('App', AppSchema)
 export const AppModel = connectionDefault.model('App', AppSchema)
+
+async function appChangeStreamListener() {
+  const appChangeStream = AppModel.watch()
+
+  appChangeStream.on('change', async change => {
+    loggers.info('Change event received')
+
+    try {
+      const response = await axios.get('/importmaps', {})
+      loggers.info('Importmaps API request successful')
+
+    } catch (e) {
+      loggers.error('Error making Importmaps API request', e);
+    }
+  })
+}
+
+appChangeStreamListener()
